+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for custom types with `from_string()` class methods
+- Python 3.14 support in CI testing matrix
+- Reserved `__cheap_settings__` attribute namespace for future configuration
+
+### Changed
+- Extended type documentation with custom type examples
+
 ## [1.3.0] - 2025-10-18
 
 ### Added